agentic_unibg/server/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# LangChain
langchain>=0.1.0
langchain-community>=0.0.16
langchain-core>=0.1.40
langchain-groq>=0.1.0
langgraph>=0.0.20

# Vector stores and embeddings
chromadb>=0.4.22
faiss-cpu>=1.7.4

# MongoDB
motor>=3.3.0
pymongo>=4.6.0

# Authentication
bcrypt>=4.1.0

# Additional dependencies
requests>=2.31.0
aiohttp>=3.9.1
python-multipart>=0.0.6
